Support slice config references. Permit build.ref to be optional in the deployment config

+++ prepareRelease.sh
@@ -8,6 +8,15 @@ trap 'exit ${RESULT:-0}' EXIT SIGHUP SIGINT SIGTERM
 DETAIL_MESSAGE="deployment=${DEPLOYMENT_TAG}, ${DETAIL_MESSAGE}"
 echo "DETAIL_MESSAGE=${DETAIL_MESSAGE}" >> ${WORKSPACE}/context.properties
 
+# Determine the slice or slices to process
+# SLICE_LIST and BUILD_SLICE may already be defined in which case SLICE and SLICES
+# are ignored.
+SLICE_LIST="${SLICE_LIST:-$SLICE}"
+SLICE_LIST="${SLICE_LIST:-$SLICES}"  
+BUILD_SLICE="${BUILD_SLICE:-$SLICE}"
+SLICE_ARRAY=($SLICE_LIST)
+BUILD_SLICE="${BUILD_SLICE:-${SLICE_ARRAY[0]}}"
+
 # Process the config repo
 cd ${WORKSPACE}/${OAID}/config/${PROJECT}
 
@@ -15,27 +24,33 @@ cd ${WORKSPACE}/${OAID}/config/${PROJECT}
 git config user.name  "${GIT_USER}"
 git config user.email "${GIT_EMAIL}"
 
-# Ensure build.ref aligns with the requested code tag
-BUILD_REFERENCE=$(echo -n "${CODE_COMMIT} ${CODE_TAG}")
-if [[ "$(cat deployments/${SEGMENT}/${BUILD_SLICE}/build.ref)" != "${BUILD_REFERENCE}" ]]; then
-	echo -n "${BUILD_REFERENCE}" > deployments/${SEGMENT}/${BUILD_SLICE}/build.ref
-	git add *; git commit -m "${DETAIL_MESSAGE}"
-	RESULT=$?
-	if [[ ${RESULT} -ne 0 ]]; then
-		echo "Can't commit the build reference to the config repo, exiting..."
-		exit
-	fi
+# Check for a deploy config reference
+if [[ -f deployments/${SEGMENT}/${BUILD_SLICE}/slice.ref ]]; then
+    BUILD_SLICE="$(cat deployments/${SEGMENT}/${BUILD_SLICE}/slice.ref)"
 fi
 
-# Generate the application level templates
+# Ensure build.ref (if present) aligns with the requested code tag
+if [[ -f deployments/${SEGMENT}/${BUILD_SLICE}/build.ref ]]; then
+    BUILD_REFERENCE=$(echo -n "${CODE_COMMIT} ${CODE_TAG}")
+    if [[ "$(cat deployments/${SEGMENT}/${BUILD_SLICE}/build.ref)" != "${BUILD_REFERENCE}" ]]; then
+        echo -n "${BUILD_REFERENCE}" > deployments/${SEGMENT}/${BUILD_SLICE}/build.ref
+        git add *; git commit -m "${DETAIL_MESSAGE}"
+        RESULT=$?
+        if [[ ${RESULT} -ne 0 ]]; then
+            echo "Can't commit the build reference to the config repo, exiting..."
+            exit
+        fi
+    fi
+fi
 BIN_DIR="${WORKSPACE}/${OAID}/config/bin"
 cd solutions/${SEGMENT}
 
-for SLICE in ${SLICE_LIST}; do
-	${BIN_DIR}/createApplicationTemplate.sh -c ${DEPLOYMENT_TAG} -s ${SLICE}
+for CURRENT_SLICE in ${SLICE_LIST}; do
+    # Generate the application level template
+	${BIN_DIR}/createApplicationTemplate.sh -c ${DEPLOYMENT_TAG} -s ${CURRENT_SLICE}
 	RESULT=$?
 	if [[ ${RESULT} -ne 0 ]]; then
- 		echo "Can't generate the template for the ${SLICE} application slice, exiting..."
+ 		echo "Can't generate the template for the ${CURRENT_SLICE} application slice, exiting..."
  		exit
 	fi
 done
